requirements.txt
# Core dependencies for CV-RAG project
# Python 3.11+ recommended

# Text processing and chunking
langchain==0.3.7
langchain-text-splitters==0.3.2

# Embeddings (local model)
sentence-transformers==3.3.1

# Database connectivity
psycopg2-binary==2.9.10
pgvector==0.3.6

# Environment variables
python-dotenv==1.0.1

# API requests (for n8n webhooks)
requests==2.32.3

# Streamlit frontend
streamlit==1.40.2

# Data handling
pandas==2.2.3

# Optional: For future API-based embeddings
# openai==1.57.4
